* src/regex.c (trim):         * src/log-common.c (format_syslog_header):         cast to int when calling isalnum and isspace.
git-svn-id: file:///home/yoann/dev/prelude/git/nok/SVN/trunk/prelude-lml@1444 09c5ec92-17d4-0310-903a-819935f44dba

+++ src/log-common.c
@@ -50,7 +50,7 @@ static int format_syslog_header(const char *buf, struct timeval *tv, char host[2
          */
         for ( i = 0; i < strlen(tag); i++ ) {
                 
-                if ( ! isalnum(tag[i]) ) {
+                if ( ! isalnum((int) tag[i]) ) {
                         tag[i] = '\0';
                         break;
                 }
